Subclass YTM Track from YT Track
So apparently songs on YouTube music are there on Youtube too, with the exact same video id. 
In fact, Youtube Music is just a layer on top of Youtube. 

Subclassing YTM Track from YT Track gives it the thumbnail property. It is used to get the thumbnail of the video